Add files via upload
Usage method: Open it in PyCharm or other Python-supported compilers, click Run, and you will see the window displaying your mouse coordinates in real time. Move your mouse to the area you want to move to, then left-click the mouse, and the (horizontal, vertical) coordinates of this location will be copied to your clipboard. Just press Ctrl+V to paste.